[#13][#22] Move test case
