+++ app/src/form/support/__test__/FieldGroupHeader.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from 'react-native-testing-library';
-import { FieldGroupHeader } from '../../../src/form/support';
+import FieldGroupHeader from '../FieldGroupHeader';
 
 describe('FieldGroupHeader component', () => {
   it('renders name and description correctly', () => {
